+++ jautograd-compute/src/main/java/com/ashutoshwad/utils/jautograd/MinBackwardOperation.java
@@ -31,7 +31,7 @@ class MinBackwardOperation extends BackwardComputeOperation {
         final double revisedGradient = resultGradient / counter;
 
         for (int[]pos : minPositions) {
-            source.accumulateGradient(pos[0], pos[1], revisedGradient);
+            source.accumulateGradient(pos[0], pos[1], (float)revisedGradient);
         }
     }
 }
